+++ common-libs/common/test/factory/didFactory.ts
@@ -1,8 +1,9 @@
 import base64url from 'base64url'
 const cryptoRandomString = require('crypto-random-string')
+import { resolveKeyDIDWithParams } from '../../src/services/DidDocumentService/KeyDidDocumentLocalResolver'
 
 import { randomBytes } from '../../src/shared/randomBytes'
-import { KeysService, DidDocumentService, DidResolver } from '../../'
+import { KeysService, DidDocumentService, DidResolver } from '../../src/index'
 
 const elemDidForLocalResolving =
   'did:elem:EiA8XCSERPjEQQJkNz55d_UZDl3_uBNFDDaeowfY7-QrPQ;elem:' +
@@ -183,6 +184,20 @@ export const generateTestDIDs = async () => {
 
   const elemBBSPublicKey = KeysService.getPublicKey(elemBBSSeedHex, 'elem').toString('hex')
 
+  const keySeed = await randomBytes(32)
+  const keySeedHex = keySeed.toString('hex')
+  const keySeedWithMethod = `${keySeedHex}++${'key'}`
+  const keyPasswordBuffer = KeysService.normalizePassword(password)
+  const keyEncryptedSeed = await KeysService.encryptSeed(keySeedWithMethod, keyPasswordBuffer)
+
+  keysService = new KeysService(keyEncryptedSeed, password)
+
+  didDocumentService = DidDocumentService.createDidDocumentService(keysService)
+  const keyDidDocument = await didDocumentService.getDidDocument(didResolverMock)
+  const keyDid = await didDocumentService.getMyDid()
+  const keyPublicKey = KeysService.getPublicKey(keySeedHex, 'elem').toString('hex')
+  const didDocumentJWK = await resolveKeyDIDWithParams(keyDid, true)
+
   return {
     password,
     jolo: {
@@ -228,6 +243,15 @@ export const generateTestDIDs = async () => {
       didDocument: webRSADidDocument,
       publicKey: webRSAPublicKey,
       publicRSAKey: webRSAPublicKeyRSA,
+    },
+    key: {
+      seed: keySeed,
+      encryptedSeed: keyEncryptedSeed,
+      seedHex: keySeedHex,
+      did: keyDid,
+      didDocument: keyDidDocument,
+      publicKey: keyPublicKey,
+      didDocumentJWK: didDocumentJWK,
     },
     elemWithRSA: {
       seed: elemRSASeed,