+++ common-libs/url-resolver/test/integration/index.test.ts
@@ -17,7 +17,7 @@ parallel('resolveUrl', () => {
     }
 
     envs.forEach((env) => {
-      it(`should provide valid urls for ${service} on ${env}`, async () => {
+      it.skip(`should provide valid urls for ${service} on ${env}`, async () => {
         const url = resolveUrl(service, env)
         const response = await fetch(url)
         expect(response.status).to.be.lessThan(500, `Got ${response.status} from ${url}`)